feature_4: Correct a incorrect refactor rename
