+++ src/main/java/com/ing/interview/api/rest/CarController.java
@@ -20,7 +20,7 @@ import javax.validation.Valid;
 import java.util.HashMap;
 import java.util.Map;
 
-import static com.ing.interview.utilsTest.ConstantsUtils.*;
+import static com.ing.interview.utils.ConstantsUtils.*;
 import static java.util.Optional.ofNullable;
 
 @RestController
