+++ src/test/java/com/ing/interview/api/rest/subordinated/webservice/configuration/CarAvailabilityServiceConfigurationTest.java
@@ -2,8 +2,8 @@ package com.ing.interview.api.rest.subordinated.webservice.configuration;
 
 import org.junit.jupiter.api.Test;
 
-import static com.ing.interview.utilsTest.ConstantsUtils.ING_STOCK_SERVICE;
-import static com.ing.interview.utilsTest.ConstantsUtils.ING_STOCK_SERVICE_PATH_TIMEOUT;
+import static com.ing.interview.utils.ConstantsUtils.ING_STOCK_SERVICE;
+import static com.ing.interview.utils.ConstantsUtils.ING_STOCK_SERVICE_PATH_TIMEOUT;
 import static org.junit.jupiter.api.Assertions.*;
 
 class CarAvailabilityServiceConfigurationTest {
